--- src/main/java/com/example/secondhand/domain/user/status/AuthHeader.java
@@ -1,7 +1,0 @@
-package com.example.secondhand.domain.user.status;
-
-public class AuthHeader {
-
-    public static final String AUTHORIZATION_HEADER = "Authorization";
-
-}
