+++ src/main/java/com/example/secondhand/global/status/Header.java
@@ -1,0 +1,7 @@
+package com.example.secondhand.global.status;
+
+public class Header {
+
+    public static final String AUTHORIZATION_HEADER = "Authorization";
+
+}
